LibWeb: Change the type of MouseEvent members to double
These are defined as integers only in the legacy specification, the
new one defines these as doubles.

+++ Userland/Libraries/LibWeb/UIEvents/MouseEvent.cpp
@@ -10,7 +10,7 @@
 
 namespace Web::UIEvents {
 
-MouseEvent::MouseEvent(const FlyString& event_name, i32 offset_x, i32 offset_y, i32 client_x, i32 client_y)
+MouseEvent::MouseEvent(const FlyString& event_name, double offset_x, double offset_y, double client_x, double client_y)
     : UIEvent(event_name)
     , m_offset_x(offset_x)
     , m_offset_y(offset_y)
